Fix for https://github.com/nanoko-project/maven-play2-plugin/issues/42 - Plugin publishes wrong application package, if it is not the last file in the OS directory listing.
Play2PackageMojo.moveApplicationPackageToTarget()  has been modified to take care of this. If the target/scala-*/ folder contains a file ending with -<version>.jar, then that is assumed to be the package artifact. Otherwise it takes the last file in the listing (as before), but ignoring a couple more file formats known to be secondary artifacts of the play build.